web/back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
ultralytics==8.0.200
torch==2.1.0
torchvision==0.16.0